+++ tests/Unit/Transform/ToStringTest.php
@@ -7,7 +7,10 @@ namespace PlainDataTransformerTests\Unit\Transform;
 use PHPUnit\Framework\TestCase;
 use PlainDataTransformer\Transform;
 use PlainDataTransformerTests\Variant\Sample;
+use PlainDataTransformerTests\Variant\SampleBackedEnum;
+use PlainDataTransformerTests\Variant\SampleBackedIntEnum;
 use PlainDataTransformerTests\Variant\SampleOne;
+use PlainDataTransformerTests\Variant\SamplePureEnum;
 use PlainDataTransformerTests\Variant\SampleTwo;
 
 class ToStringTest extends TestCase
@@ -36,6 +39,9 @@ class ToStringTest extends TestCase
     {
         $this->assertSame('', Transform::toString(new Sample()));
         $this->assertSame('sample', Transform::toString(new SampleOne()));
+        $this->assertSame('ACTIVE', Transform::toString(SamplePureEnum::ACTIVE));
+        $this->assertSame('type_a', Transform::toString(SampleBackedEnum::TYPE_A));
+        $this->assertSame('1', Transform::toString(SampleBackedIntEnum::TYPE_ONE));
         $this->assertSame('0.1', Transform::toString(0.1));
         $this->assertSame('1', Transform::toString(1));
         $this->assertSame('true', Transform::toString(true));