+++ tests/Variant/SampleBackedIntEnum.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PlainDataTransformerTests\Variant;
+
+enum SampleBackedIntEnum: int
+{
+    case TYPE_ONE = 1;
+    case TYPE_TWO = 2;
+    case TYPE_THREE = 3;
+}